+++ R/get_label.R
@@ -57,9 +57,7 @@
 #'         main = get_label(efc$e42dep))
 #'
 #' # get labels from multiple variables
-#' get_label(list(efc$e42dep,
-#'                efc$e16sex,
-#'                efc$e15relat))
+#' get_label(list(efc$e42dep, efc$e16sex, efc$e15relat))
 #'
 #' @export
 get_label <- function(x, def.value = NULL) {