+++ R/rmse.R
@@ -41,7 +41,7 @@
 #'
 #' # normalized RMSE
 #' library(nlme)
-#' fit <- lme(distance ~ age, data = Orthodont) # random is ~ age
+#' fit <- lme(distance ~ age, data = Orthodont)
 #' rmse(fit, normalized = TRUE)
 #'
 #' @importFrom stats residuals