+++ R/to_factor.R
@@ -50,10 +50,8 @@
 #'
 #' data(efc)
 #' # create parially labelled vector
-#' x <- set_labels(efc$e42dep,
-#'                 c(`1` = "independent",
-#'                   `4` = "severe dependency",
-#'                   `9` = "missing value"))
+#' x <- set_labels(efc$e42dep, c(`1` = "independent", `4` = "severe dependency",
+#'                               `9` = "missing value"))
 #'
 #' # only copy existing value labels
 #' to_factor(x)