+++ R/to_long.R
@@ -117,10 +117,7 @@ to_long <- function(data, keys, values, ..., labels = NULL, recode.key = FALSE) 
     # remove those columns that should not be gathered
     tmp <- data[, -match(remove_cols, colnames(data))]
     # gather data frame
-    tmp <- suppressWarnings(tidyr::gather_(tmp,
-                                           keys[i],
-                                           values[i],
-                                           data_cols[[i]]))
+    tmp <- suppressWarnings(tidyr::gather_(tmp, keys[i], values[i], data_cols[[i]]))
     # need to recode key-value?
     if (recode.key)
       tmp[[keys[i]]] <- sort(to_value(tmp[[keys[i]]], keep.labels = FALSE))