+++ NEWS.md
@@ -1,8 +1,10 @@
 # sjmisc 1.8
 
 ## New functions
+
 * `samplesize.lmm` to compute approximated sample size for two-level linear mixed models.
 * `deff` to compute the design effect for two-level linear mixed models.
+* `get_re_var` to get specific components of the random effect variances from mixed models.
 
 ## Changes to functions